+++ scripts/hooks.js
@@ -50,6 +50,7 @@ export function registerHooks() {
     Hooks.on('midi-qol.premades.postWaitForAttackRoll', midiEvents.postAttackRoll);
     Hooks.on('midi-qol.premades.postAttackRollComplete', midiEvents.attackRollComplete);
     Hooks.on('midi-qol.premades.preDamageRollComplete', midiEvents.damageRollComplete);
+    Hooks.on('midi-qol.premades.preUtilityRollComplete', midiEvents.utilityRollComplete);
     Hooks.on('midi-qol.premades.postSavesComplete', midiEvents.savesComplete);
     Hooks.on('midi-qol.preTargetDamageApplication', midiEvents.preTargetDamageApplication);
     Hooks.on('midi-qol.premades.postRollFinished', midiEvents.rollFinished);
@@ -155,6 +156,7 @@ export function registerHooks() {
         Hooks.on('canvasReady', auras.canvasReady);
         Hooks.on('createItem', itemEvent.created);
         Hooks.on('deleteItem', itemEvent.deleted);
+        Hooks.on('updateItem', itemEvent.updated);
 
         // Scene & Compendium medkit
         Hooks.on('getHeaderControlsSceneConfig', appendHeaderControl);