+++ scripts/legacyMacros.js
@@ -483,6 +483,7 @@ export {telepathicInsight} from './macros/2014/raceFeatures/verdan/telepathicIns
 export {vampiricBite} from './macros/2014/raceFeatures/dhampir/vampiricBite.js';
 export {wardersIntuition} from './macros/2014/raceFeatures/dwarf/markOfWarding/wardersIntuition.js';
 export {windwrightsIntuition} from './macros/2014/raceFeatures/halfElf/markOfStorm/windwrightsIntuition.js';
+export {huntersIntuition} from './macros/2014/raceFeatures/human/markOfFinding/huntersIntuition.js';
 // Monster Features
 export {armyArcana} from './macros/2014/monsterFeatures/fey/hobgoblinDevastator/armyArcana.js';
 export {causticMist} from './macros/2014/monsterFeatures/monstrosity/hoardMimic/causticMist.js';
@@ -646,6 +647,11 @@ export {unarmedStrike} from './macros/2014/items/misc/unarmedStrike.js';
 export {wandOfTheWarMage, wandOfTheWarMage1, wandOfTheWarMage2, wandOfTheWarMage3} from './macros/2014/items/trinket/wandOfTheWarMage.js';
 export {wingsOfFlying, wingsOfFlyingEffect, wingsOfFlyingRecharge} from './macros/2014/items/clothing/wingsOfFlying.js';
 export {wrapsOfDyamak, wrapsOfDyamak0, wrapsOfDyamak1, wrapsOfDyamak2, crimsonMist} from './macros/2014/items/trinket/wrapsOfDyamak.js';
+export {robeOfUsefulItems, robeOfUsefulItemsSetupPatch, robeOfUsefulItemsPatch} from './macros/2014/items/clothing/robeOfUsefulItems.js';
+export {ruinsWakeD, ruinsWakeA, ruinsWakeE} from './macros/2014/items/weapons/ruinsWake.js';
+export {charlatansDie} from './macros/2014/items/trinket/charlatansDie.js';
+export {deckOfMiscellany, deckOfMiscellanySetup, miscellanyCard} from './macros/2014/items/trinket/deckOfMiscellany.js';
+export {deckOfWonder, deckOfWonderDestinyEffect, deckOfWonderIsolationEffect, deckOfWonderJusticeEffect, deckOfWonderVultureLostEffect} from './macros/2014/items/trinket/deckOfWonder.js';
 //BG3 Weapon Actions
 export {backbreaker, braceMelee, braceMeleeDamage, cleave, concussiveSmash, maimingStrike, disarmingStrike, flourish, heartstopper, lacerate, piercingStrike, piercingStrikeEffect, pommelStrike, prepare, rush, tenacity, topple, weakeningStrike, braceRanged, braceRangedDamage, hamstringShot, mobileShot, piercingShot} from './macros/2014/homebrew/bg3WeaponActions.js';
 //Other