+++ scripts/lib/utilities/socketUtils.js
@@ -2,12 +2,10 @@ import {socket, sockets} from '../sockets.js';
 import {genericUtils} from '../../utils.js';
 function gmID() {
     let gmID = game.settings.get('chris-premades', 'gmID');
-    let preferredGM = game.settings.get('midi-qol', 'PreferredGM');  
-    if (preferredGM !== 'none') {
-        let preferredGM = game.users.get(preferredGM);
-        if (preferredGM?.active) {
-            gmID = preferredGM.id;
-        }
+    let preferredGMId = game.settings.get('midi-qol', 'PreferredGM');  
+    if (preferredGMId !== '') {
+        let preferredGM = game.users.get(preferredGMId);
+        if (preferredGM?.active) gmID = preferredGM.id;
     }
     return gmID;
 }